feat(assets): implement useAssetsBalance hook for fetching asset balances
